Added the forced_primordial_mmw member to the chemistry_data struct which is passed as priForcedMMW to cool1d_multi_g, solve_rate_cool_g, and calc_temp1d_cloudy_g.
If forced_primordial_mmw>0 when primordial_chemistry=0, the mean molecular mass of the primordial gas is set to the value of forced_primordial_mmw (the tabulated mean molecular weights are ignored). When forced_primordial_mmw<=0, then grackle's behavior is unchanged (forced_primordial_mmw is 0 by default so that the default behavior of grackle is unchanged). The call to initialize_chemistry_data will fail and report an error if forced_primordial_mmw>0 and primordial_chemistry > 0.