Added a unit test for the new forced_primordial_mmw member of the chemistry_data struct.
For a given random set of Temperatures, density, and mmw (mean molecular weight), this test computes the corresponding internal energies. Next, a chemistry_data instance is created and the forced_primordial_mmw property set to the given mmw. Then, the chemistry_data instance is used to initialize a FluidContainer object in order to compute the temperature from the given density and previously computed internal energy. Finally it uses the internal energies and newly computed temperature to determine the mean molecular weight (used internally by the C/Fortran functions), and checks that it is the same as the given mmw (to within floating point precision).