Modify unit tests for ``cool1d_cloudy_g`` and ``calc_temp1d_cloudy_g`` so that the legacy fortran versions now call the C-versions of the interpolation routines.
Before this commit:
- the legacy fortran versions of ``cool1d_cloudy_g`` and ``calc_temp1d_cloudy_g`` called the legacy-fortran versions of the interpolation routines (which are no longer included in the Grackle library)
- the newer C-versions of ``cool1d_cloudy_g`` and ``calc_temp1d_cloudy_g`` called the newer C-versions of the interpolation routines (that are included in the Grackle library)

This choice was fine; all of the routines produced consistent results. But, it might produce some unexpected issues in the future if we ever optimize the interpolation routines and tolerate round-off error compared to the legacy fortran routines.

After this commit, the Fortran-versions and the C-versions of ``cool1d_cloudy_g`` and ``calc_temp1d_cloudy_g`` all use the C-version of the interpolation routines.

(NOTE: This has no impact on the unit tests that directly test the interpolation routines. Within that test, the new C-versions are ALWAYS tested against the legacy Fortran versions)